requirements.txt
mysql-connector-python==8.4.0
numpy==1.24.4
pandas==2.0.3